Player search working!

+++ dota/src/main/java/com/hatfat/dota/model/user/SteamUser.java
@@ -251,7 +251,7 @@ public class SteamUser {
 
     @Override
     public int hashCode() {
-        return Integer.parseInt(steamId);
+        return (int) Long.parseLong(steamId);
     }
 
     public static Comparator<SteamUser> getComparator() {
